"Fix: Estabilización Backend Admin y mejoras UI Frontend
- BACKEND:
  - Revertido modelo Usuario a 'passwordHash' para compatibilidad BBDD.
  - AdminController: Inyección de PasswordEncoder y borrado en cascada (Transactional).
  - Añadidos métodos en Repositorios para búsquedas y borrados por entidad.
  - Corrección en búsqueda de partidos por equipo.

- FRONTEND:
  - AdminDashboard: Lógica de borrado de usuarios y navegación con parámetros.
  - AdminDashboard: Rediseño de pestaña 'Competición' con tarjetas visuales.
  - CalendarPage: Soporte para filtrar eventos por 'teamId' vía URL (modo Admi

+++ src/backend-tfg/backend-tfg/src/main/java/com/DAMUnitedFC/backend_tfg/controller/JugadorController.java
@@ -80,7 +80,7 @@ public class JugadorController {
 
     // --- DASHBOARD JUGADOR ---
     @GetMapping("/usuario/{idUsuario}/equipo")
-    public ResponseEntity<?> getEquipoDelJugador(@PathVariable Long idUsuario) {
+    public ResponseEntity<?> getEquipoDelJugador(@PathVariable Integer idUsuario) {
         return repo.findByUsuario_IdUsuario(idUsuario)
                 .map(jugador -> {
                     if (jugador.getEquipoPrincipal() != null) {
